fix: API 요청 시 토큰 확인 로직 수정
- 권한 오류 발생 시 토큰 발급 시도 후 실패 시 이동, 나머지 오류는 그대로 전달
- /live-map의 header 타이틀을 모임 타이틀로 변경
- 인터넷 에러 발생 시, 홈으로 돌아가는 버튼 변경

+++ frontend/src/api/index.js
@@ -28,6 +28,7 @@ function apiInstance() {
       return response;
     },
     async function (error) {
+      console.log(error)
       var result = null;
       // 권한 오류인 경우, access_token 재발급 시도
       if (error.response.data.status == 401) {
@@ -46,11 +47,15 @@ function apiInstance() {
             localStorage.setItem("Authorization", response.data);
             result = await instance(error.config);
           })
-          .catch((error) => {
-            error;
-            localStorage.clear();
-            window.location.href = "/login";
-            return true;
+          .catch(async (error) => {
+            const data = error.response.data;
+            // access token 재발급 불가 또는 존재하지 않는 회원인 경우
+            if (data.status == 500 || (data.status == 404 && data.message == "member not found.")) {
+              localStorage.clear();
+              window.location.href = "/login";
+            } else {
+              result = await Promise.reject(error);
+            }
           });
       } else {
         // 그 외의 경우는 오류 그대로 전달
